FundSquare - Approach 2/requirements.txt
selenium==4.15.0
beautifulsoup4==4.12.2
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
openpyxl==3.1.2
webdriver-manager==4.0.1
fake-useragent==1.4.0
